pyproject.toml
[project]
name = "youtube-check-8k-agent"
version = "0.1.0"
description = "YouTube Studio 8K checker with LangChain + Ollama"
requires-python = ">=3.10"
dependencies = [
  "apscheduler==3.10.4",
  "fastapi==0.115.5",
  "httpx==0.27.2",
  "langchain==1.2.10",
  "langchain-community==0.4.1",
  "langchain-ollama>=1.0.1",
  "playwright==1.58.0",
  "pydantic==2.9.2",
  "python-dotenv==1.0.1",
  "tzlocal==5.2",
  "uvicorn==0.30.6",
]

[dependency-groups]
dev = []
